+++ Rome2Rio/Rome2Rio/R2RStringFormatter.h
@@ -23,7 +23,7 @@
 +(NSString *) formatDurationZeroPadded: (float) minutes;
 +(NSString *) formatFrequency: (float) frequency;
 +(NSString *) formatDistance: (float) distance isImperial: (bool) isImperial;
-+(NSString *) formatDays: (int) days;
++(NSString *) formatDays: (NSInteger) days;
 
 +(NSString *) capitaliseFirstLetter: (NSString *) string;
